Concatenate Text (add to make_dataframes) and first version of CV and SGD-regressor (ElasticNet)

+++ data_processing/make_dataframes.py
@@ -21,6 +21,7 @@ import pandas as pd
 import sys 
 import re
 import json
+import os
 
 if len(sys.argv) != 2:
     print("Usage: python make_dataframes.py ~/Path/To/Data/")
@@ -63,34 +64,46 @@ def extract_web_addresses(df):
         df.at[index, "fb_status_msg"] = new_text
     return df, link_dict
 
-dataframe, link_dict = extract_web_addresses(full_data) # replace http(s):// links with UUIDs
-
 link_path = source_data_dir + '/' + "web_links.json"
-with open(link_path, 'w', encoding='utf-8') as fo:
-    json.dump(link_dict, fo, ensure_ascii=False, indent=4)
+dataframe, link_dict = extract_web_addresses(full_data) # replace http(s):// links with UUIDs
+       
+if not os.path.isfile(link_path):
+    with open(link_path, 'w', encoding='utf-8') as fo:
+        json.dump(link_dict, fo, ensure_ascii=False, indent=4)
 
 geo_cols = ["country_current", "county code", "hometown", "zipcode"]
 demographics = ["politics" + str(int_val) + "way" for int_val in [10,7,4] ] + \
         ["age", "concentrations", "degrees", "significant_other_id", "relation_status", \
          "religion_now", "politics_new", "sex", "gender"]
-
-
 demo_race = ["race_" + col for col in ["black", "eastasian", "latino", "middleeast", "nativeamerican", "southasian", "white", "other", "decline"] ]
-
 
 demographics = geo_cols + demographics + demo_race
 predictors = ["fb_status_msg"]
 outcomes = [col for col in full_data.columns.values if (col.startswith("MFQ") and col not in ["MFQ_HFminsIAP", "MFQ_BLANKS_perc", "MFQ_failed"]) or col == "mfq_unclear"]
 
 full_selected = dataframe.loc[:, ["userid"] + demographics + predictors + outcomes]
-#text_plus_outcomes = full_data.loc[:, ["userid"] + predictors + outcomes]
-#demo_plus_outcomes = full_data.loc[:, ["userid"] + demographics + outcomes]
 
 dest = source_data_dir + '/' + "full_dataframe.pkl"
 full_selected.to_pickle(dest)
 
-#dest = source_data_dir + '/' + "text_and_outcomes.pkl"
-#text_plus_outcomes.to_pickle(dest)
+# make concatenated_dataframe
+users = dict()
+for index, row in full_selected.iterrows():
+    userid = row["userid"]
+    if userid not in users:
+        users[userid] = [row.to_dict()]
+    else:
+        users[userid].append(row.to_dict())
 
-#dest = source_data_dir + '/' + "demographics_and_outcomes.pkl"
-#demo_plus_outcomes.to_pickle(dest)
+new_df = list()
+for userid in users:
+    texts = list()
+    for row in users[userid]:
+        texts.append(row["fb_status_msg"])
+    new_text = "\n".join(texts)
+    users[userid][0]["fb_status_msg"] = new_text
+    new_df.append(users[userid][0])
+
+concat_df = pd.DataFrame(new_df)
+concat_df.to_pickle(source_data_dir + '/' + "concat_df.pkl")
+
